Group related business logic into modules
Introduce TaskOrderErrorProvider module that is used to
get error messages and payloads for all types of errors that the
TaskOrderService can return

Introduce TaskOrderResultEncoder module that is used to format the
response in an appropriate manner. Plain text responses do not get a
payload, for example, since they are meant to run via bash. Instead a
script that echoes the error is returned.